test(postgres): Add unit tests with 60.9% coverage
- 57 tests covering CRUD, I/O, move/remove, symlinks, capabilities
- Shared PostgreSQL container via testcontainers-go (~3.5s execution)
- Fixed file_type constraint (0-6) and move operation type cast
- Updated docs: CONFIGURATION, ARCHITECTURE, status tracking

+++ pkg/store/metadata/postgres/move.go
@@ -255,7 +255,7 @@ func (s *PostgresMetadataStore) Move(
 
 		updateDescendantsQuery := `
 			UPDATE files
-			SET path = $1 || SUBSTRING(path FROM $2),
+			SET path = $1 || SUBSTRING(path FROM $2::INTEGER),
 			    ctime = $3
 			WHERE path LIKE $4 AND share_name = $5
 		`
